Update main and re-org code

+++ get_next_line_bonus/get_next_line_bonus.c
@@ -9,7 +9,7 @@
 /*   Updated: 2023/11/18 23:43:48 by ejohnson         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-#include "get_next_line_bonus.h"
+#include "../get_next_line_utils.h"
 
 char	*ft_strjoin(char *s1, char const *s2)
 {
@@ -35,8 +35,8 @@ char	*ft_strjoin(char *s1, char const *s2)
 
 static char	*read_backup(int fd, char *buff, char *backup)
 {
-	int	bytes;
-	int	sentry;
+	long	bytes;
+	int sentry;
 
 	bytes = 1;
 	sentry = 1;
